# modified freelist_hmap/hashmapGetFreePageIDs with better performance